Add ManageRace Schedule

+++ DakarRally.Net-dusanj.Domain/Entity/Car.cs
@@ -6,19 +6,19 @@ namespace DakarRally.Net_dusanj.Domain.Entity
     [Table("Cars")]
     public class Car : Vehicle
     {
-        public const int MaxSpeedSport = 140;
+        public const decimal MaxSpeedSport = 140;
 
-        public const int MaxSpeedTerrain = 100;
+        public const decimal MaxSpeedTerrain = 100;
 
-        public const int RepairmentDuration = 5;
+        public const decimal RepairmentDuration = 5;
 
-        public const int MalfunctionProbabilitySportLight = 12;
+        public const decimal MalfunctionProbabilitySportLight = 12;
 
-        public const int MalfunctionProbabilitySportHeavy = 2;
+        public const decimal MalfunctionProbabilitySportHeavy = 2;
 
-        public const int MalfunctionProbabilityTerrainLight = 3;
+        public const decimal MalfunctionProbabilityTerrainLight = 3;
 
-        public const int MalfunctionProbabilityTerrainHeavy = 1;
+        public const decimal MalfunctionProbabilityTerrainHeavy = 1;
 
         public CarTypeEnum CarType { get; set; }
     }
