+++ DakarRally.Net-dusanj.Domain/Entity/Truck.cs
@@ -5,12 +5,12 @@ namespace DakarRally.Net_dusanj.Domain.Entity
     [Table("Trucks")]
     public class Truck : Vehicle
     {
-        public const int MaxSpeed = 140;
+        public const decimal MaxSpeed = 140;
 
-        public const int RepairmentDuration = 7;
+        public const decimal RepairmentDuration = 7;
 
-        public const int MalfunctionProbabilityLight = 6;
+        public const decimal MalfunctionProbabilityLight = 6;
 
-        public const int MalfunctionProbabilityHeavy = 4;
+        public const decimal MalfunctionProbabilityHeavy = 4;
     }
 }
